data type change

+++ test_dags/bigquery/alloc.sql
@@ -1,5 +1,5 @@
 create table `dmgcp-ingestion-poc.airflow_test.test_dag` (
 name string,
 address string,
-phone integer
+phone int64
 )
